Separate out the contact form
For improving the home component.
