+++ src/templates/Home/ContactForm/index.ts
@@ -1,0 +1,1 @@
+export {ContactForm} from './ContactForm'
